// src/app/models/commons/Relations.ts
import {Item} from "./Item";

export class Relations {
  apartmentConditions: Item[];
  cities: Item[];
  apartmentFurniture: Item[];
  apartmentFacilities: Item[];
  apartmentBathroomTypes: Item[];
  apartmentFurnitureStatuses: Item[];
  apartmentBathrooms: Item[];
  apartmentSecurities: Item[];
  windowDirections: Item[];
  apartmentFor: Item[];
  ad_gender_types: Item[];
}